backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pyjwt
httpx
python-telegram-bot
pydantic-settings